+++ tests/playwright/sanity/modules/home/data/home-screen.mock.ts
