+++ tests/playwright/sanity/modules/home/home-screen.helper.ts
@@ -1,0 +1,28 @@
+import { type Page } from '@playwright/test';
+import homeScreenMockData from './data/home-screen.mock';
+
+export type LicenseType = 'free' | 'pro' | 'one';
+
+export const transformMockDataByLicense = ( licenseType: LicenseType ) => {
+	const topItem = homeScreenMockData.top_with_licences.find( ( item ) => item.license.includes( licenseType ) )!;
+	const getStartedItem = homeScreenMockData.get_started.find( ( item ) => item.license.includes( licenseType ) )!;
+	const sidebarPromotionItem = homeScreenMockData.sidebar_promotion_variants.find(
+		( item ) => 'true' === item.is_enabled && item.license.includes( licenseType ),
+	);
+
+	return {
+		top_with_licences: topItem,
+		get_started: getStartedItem,
+		add_ons: homeScreenMockData.add_ons,
+		sidebar_upgrade: homeScreenMockData.sidebar_upgrade,
+		...( sidebarPromotionItem && { sidebar_promotion_variants: sidebarPromotionItem } ),
+		external_links: homeScreenMockData.external_links,
+	};
+};
+
+export const mockHomeScreenData = async ( page: Page, mockData: ReturnType<typeof transformMockDataByLicense> ) => {
+	await page.addInitScript( ( data ) => {
+		// @ts-expect-error - window.elementorHomeScreenData is dynamically assigned
+		window.elementorHomeScreenData = data;
+	}, mockData );
+};
